+++ yudao-module-erp/yudao-module-erp-biz/src/main/java/cn/iocoder/yudao/module/erp/service/product/bo/ErpProductBO.java
@@ -49,7 +49,7 @@ public class ErpProductBO {
     /**
      * 基础重量（kg）
      */
-    private BigDecimal weight;
+    private Integer weight;
     /**
      * 系列
      */
@@ -73,15 +73,15 @@ public class ErpProductBO {
     /**
      * 基础宽度（mm）
      */
-    private BigDecimal width;
+    private Integer width;
     /**
      * 基础长度（mm）
      */
-    private BigDecimal length;
+    private Integer length;
     /**
      * 基础高度（mm）
      */
-    private BigDecimal height;
+    private Integer height;
     /**
      * 主图url
      */
